Configured state management using provider for map, re-organization of codebase, built new widget and added logic for rider UI
