+++ .packages
@@ -3,7 +3,7 @@
 # 
 # For more info see: https://dart.dev/go/dot-packages-deprecation
 # 
-# Generated by pub on 2022-03-13 13:39:40.144580.
+# Generated by pub on 2022-04-09 12:25:44.356224.
 async:file:///Applications/flutter/.pub-cache/hosted/pub.dartlang.org/async-2.8.2/lib/
 boolean_selector:file:///Applications/flutter/.pub-cache/hosted/pub.dartlang.org/boolean_selector-2.1.0/lib/
 characters:file:///Applications/flutter/.pub-cache/hosted/pub.dartlang.org/characters-1.2.0/lib/
@@ -65,6 +65,7 @@ pedantic:file:///Applications/flutter/.pub-cache/hosted/pub.dartlang.org/pedanti
 platform:file:///Applications/flutter/.pub-cache/hosted/pub.dartlang.org/platform-3.1.0/lib/
 plugin_platform_interface:file:///Applications/flutter/.pub-cache/hosted/pub.dartlang.org/plugin_platform_interface-2.1.2/lib/
 process:file:///Applications/flutter/.pub-cache/hosted/pub.dartlang.org/process-4.2.4/lib/
+progress_indicators:file:///Applications/flutter/.pub-cache/hosted/pub.dartlang.org/progress_indicators-1.0.0/lib/
 provider:file:///Applications/flutter/.pub-cache/hosted/pub.dartlang.org/provider-6.0.2/lib/
 quiver:file:///Applications/flutter/.pub-cache/hosted/pub.dartlang.org/quiver-3.0.1+1/lib/
 sanitize_html:file:///Applications/flutter/.pub-cache/hosted/pub.dartlang.org/sanitize_html-2.0.0/lib/
