+++ lib/screens/launch_screen.dart
@@ -1,4 +1,4 @@
-import 'package:carpool_app/base_screen.dart';
+import 'package:carpool_app/screens/base_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:smooth_page_indicator/smooth_page_indicator.dart';
 import 'package:provider/provider.dart';
