+++ lib/screens/signin_screen.dart
@@ -1,7 +1,7 @@
 import 'dart:collection';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'auth.dart';
+import 'package:carpool_app/services/auth_service.dart';
 
 class SignInScreen extends StatefulWidget {
   final String role;
