+++ lib/screens/signup_screen.dart
@@ -1,6 +1,6 @@
 import 'dart:collection';
 import 'package:flutter/material.dart';
-import 'package:carpool_app/auth.dart';
+import 'package:carpool_app/services/auth_service.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 
 class SignUpScreen extends StatefulWidget {
