+++ lib/services/auth_service.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:uuid/uuid.dart';
 import 'package:firebase_core/firebase_core.dart';
-import 'firebase_options.dart';
+import '../firebase_options.dart';
 import 'package:google_sign_in/google_sign_in.dart';
 import 'dart:async';
 
@@ -21,6 +21,10 @@ class AuthService {
 
   final googleSignIn = GoogleSignIn();
   String _failureReason = "None";
+
+  String getCurrentUserRole() {
+    return currentUserRole;
+  }
 
   Stream<User?> get user {
     return _auth.authStateChanges();
