+++ lib/services/map_metrics_service.dart
@@ -1,0 +1,14 @@
+  
+  Future<String>? getDistanceInMiles(var directions) async {
+    if (directions['distanceInMiles'] == "") {
+      return "0 mi";
+    }
+    return directions['distanceInMiles'];
+  }
+
+  Future<String>? getDurationToReachDestination(var directions) async {
+    if (directions['durationToReachDestination'] == "") {
+      return "0 mi";
+    }
+    return directions['durationToReachDestination'];
+  }
